Add files via upload
graphics.py contains all of the graphics and methods for manipulating graphics. GUI.py is the graphical user interface (still in progress), TUI is the textual interface (in the console), Dice.py contains the dice object along with dice functions like roll(), and Yahtzee.py contains the code to play the game.